+++ api/tests/test_views.py
@@ -18,10 +18,11 @@ class ViewsTestCase(TestCase):
         """clean up residual test files."""
         File.objects.all().delete()
         pattern = "^(?=test_file)\w+"
-        for the_file in os.listdir("media/files"):
-            # remove all the uploaded test files
-            if re.search(pattern, the_file):
-                os.remove(os.getcwd() + '/media/files/' + the_file)
+        if os.path.isdir(os.getcwd() + '/media/files/'):
+            for the_file in os.listdir("media/files"):
+                # remove all the uploaded test files
+                if re.search(pattern, the_file):
+                    os.remove(os.getcwd() + '/media/files/' + the_file)
 
     def create_file(self, filepath):
         """Create a file for testing."""
